Rename RoomListView -> SideNavView
