+++ src/components/SideNavView.tsx
